requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-ckeditor==6.5.0
django-content-editor==6.3.1
django-extensions==3.2.0
django-js-asset==2.0.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
python-dateutil==1.5
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
